app/app-requirements.txt
streamlit
geopandas
sqlalchemy
psycopg2-binary
python-dotenv
folium
streamlit-folium
